backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
pytest>=8.2.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.1
